requirements.txt
# Core requirements
streamlit==1.24.0
pandas==2.0.3
numpy==1.24.3
youtube-transcript-api==0.6.1
requests==2.31.0
nltk==3.8.1
textblob==0.17.1
wordcloud==1.9.3
matplotlib==3.7.1
python-dotenv==1.0.0
openai==1.12.0
supabase==2.3.0
yt-dlp>=2024.3.10
protobuf==3.20.3
google-api-python-client>=2.0.0

# ML requirements (install these separately)
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.0+cpu
torchaudio==2.1.0+cpu
openai-whisper==20231117
typing-extensions>=4.7.1
pydantic>=2.0.0
